requirements.txt
flask
flask-cors
tensorflow-cpu==2.12.0
pillow
numpy
gunicorn
